Move issue and redeem methods to the GiftCardService The issue and redeem methods apply to a GiftCard and not to a Transaction. This also removes the requirement of retrieving a GiftCard before a new issue or redeem action can be created.
