+++ .editorconfig
@@ -12,3 +12,6 @@ trim_trailing_whitespace = true
 
 [*.json]
 indent_size = 2
+
+[*.md]
+indent_size = 2
